Ensure the order is complete before sending to another page
If the user's order is not yet complete, we'll render a new "processing your order" view that waits for the IPN to complete the order. It refreshes every 6 seconds until the IPN succeeds.

We should eventually redirect to an error page after x attempts, or if the order has been marked as failed.